requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
distro==1.9.0
filelock==3.14.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.30.3
outcome==1.3.0.post0
pipenv==2023.12.1
platformdirs==4.2.2
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.2
selenium==4.21.0
setuptools==70.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
virtualenv==20.26.2
wsproto==1.2.0
msgraph-sdk==1.11.0